@@ -7,16 +7,29 @@ use PHPUnit\Framework\TestCase;
 
 final class WhitelistTest extends TestCase
 {
-    /** @covers \Kalessil\Composer\Plugins\ProductionDependenciesGuard\Whitelist::<public> */
-    public function testComponent()
+    public function testComponent(): void
     {
         $package = $this->createMock(CompletePackageInterface::class);
-        $package->expects($this->atLeastOnce())->method('getName')->willReturn('package1', 'package2', 'package3');
+        $package->expects($this->atLeastOnce())->method('getName')->willReturn(
+            'package1',
+            'Package2',
+            'package2',
+            'package3',
+            '...',
+            'vendor/package',
+        );
 
-        $component = new Whitelist(['', '...', 'white-list:package1', 'white-list:package2']);
+        $component = new Whitelist([
+            'package1' => [],
+            'package2' => [ 'abandoned' ],
+            '...' => [ 'description' ],
+        ]);
 
-        $this->assertTrue($component->canUse($package));
-        $this->assertTrue($component->canUse($package));
-        $this->assertFalse($component->canUse($package));
+        $this->assertTrue($component->canUse($package, 'lock-file'));
+        $this->assertTrue($component->canUse($package, 'abandoned'));
+        $this->assertFalse($component->canUse($package, 'license'));
+        $this->assertFalse($component->canUse($package, 'description'));
+        $this->assertTrue($component->canUse($package, 'description'));
+        $this->assertFalse($component->canUse($package, 'description'));
     }
 }
